moved user report api to api module in accounts
