+++ ledger/payments/api.py
@@ -1,4 +1,3 @@
-from django.db import connection
 from django.http import HttpResponse, HttpResponseRedirect
 from django.template.loader import get_template
 from django.template import TemplateDoesNotExist
@@ -18,7 +17,6 @@ from ledger.payments.facade import bpoint_facade
 from ledger.payments.reports import generate_items_csv, generate_trans_csv
 
 from ledger.accounts.models import EmailUser
-from ledger.accounts.reports import user_report
 from ledger.catalogue.models import Product
 from oscar.apps.order.models import Order
 from oscar.apps.voucher.models import Voucher
@@ -750,28 +748,6 @@ class ReportCreateView(views.APIView):
                                             ,serializer.validated_data['start'],
                                             serializer.validated_data['end'],
                                             district = serializer.validated_data['district'])
-            if report:
-                response = HttpResponse(FileWrapper(report), content_type='text/csv')
-                response['Content-Disposition'] = 'attachment; filename={}.csv'.format(filename)
-                return response
-            else:
-                raise serializers.ValidationError('No report was generated.')
-        except serializers.ValidationError:
-            raise
-        except Exception as e:
-            traceback.print_exc()
-            raise serializers.ValidationError(str(e))
-
-class UserReportView(views.APIView):
-
-    def get(self,request,format=None):
-        try:
-            http_status = status.HTTP_200_OK
-            report = None
-
-            filename = 'emailuser-report'
-            # Generate Report
-            report = user_report()
             if report:
                 response = HttpResponse(FileWrapper(report), content_type='text/csv')
                 response['Content-Disposition'] = 'attachment; filename={}.csv'.format(filename)
